requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
boltons==21.0.0
bracex==2.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
dataclasses-json==0.6.1
defusedxml==0.7.1
face==22.0.0
Faker==19.6.2
favicon==0.7.0
filelock==3.12.4
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.37
glom==22.1.0
greenlet==2.0.2
htbuilder==0.6.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.305
langsmith==0.0.41
litellm==0.1.807
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
open-interpreter==0.1.6
openai==0.28.1
packaging==23.2
pandas==2.1.1
peewee==3.16.3
Pillow==10.0.1
protobuf==4.24.3
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-lsp-jsonrpc==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
semgrep==1.42.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
st-annotated-text==4.0.1
streamlit==1.27.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
streamlit-option-menu==0.3.6
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.4.0
tokenizers==0.14.0
tokentrim==0.1.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.16
uuid==1.30
validators==0.22.0
watchdog==3.0.0
wcmatch==8.5
wcwidth==0.2.8
wget==3.2
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
loguru